[#33] - Dialog de Confirmação

+++ src/app/courses/courses/courses.component.ts
@@ -7,6 +7,7 @@ import { catchError } from 'rxjs/operators';
 import { Course } from '../model/course';
 import { CoursesService } from '../services/courses.service';
 import { ErrorDialogComponent } from './../../shared/components/error-dialog/error-dialog.component';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -55,17 +56,26 @@ export class CoursesComponent implements OnInit {
   }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course.id).subscribe(
-      () => {
-        this.refresh()
-        this._snackBar.open('Curso removido com sucesso', 'X', {
-          duration: 3000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
-        });
-      },
-      () => this.onError("Erro ao tentar remover curso")
-    );
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course.id).subscribe(
+          () => {
+            this.refresh()
+            this._snackBar.open('Curso removido com sucesso', 'X', {
+              duration: 3000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            });
+          },
+          () => this.onError("Erro ao tentar remover curso")
+        );
+      }
+    });
+
   }
 
 }
